# watsonx/tools/quotecraft-api-openapi.yaml
openapi: 3.0.0
info:
  title: QuoteCraft Procurement API
  description: API for automated BOQ and vendor quote processing
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
paths:
...
  /api/erp/create-po:
    post:
      summary: Create Purchase Order
      description: Create PO in ERP system
      operationId: createPurchaseOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - selectedVendor
                - comparisonId
              properties:
                selectedVendor:
                  type: string
                comparisonId:
                  type: string
                poData:
                  type: object
      responses:
        '200':
          description: PO created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      poNumber:
                        type: string
                      status:
                        type: string
...